Creates custom NixOS installer ISO
Defines a flake that builds a custom NixOS ISO.

This ISO clones a specified git repository to a directory within the target system during installation.
It achieves this by integrating a Calamares module that executes a script within the chroot environment.
The Calamares settings are patched to ensure this module runs before the final installation steps.